+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+
+- Allowed `PastDate` and `FutureDate` validators to use `datetime` objects
+- Date fields can use ISO 8601 dates as `data` in addition to the existing date formats
+
 ### Security
 
 ## [0.26.0](https://github.com/nationalarchives/tna-frontend-jinja/compare/v0.25.0...v0.26.0) - 2025-10-07